+++ packages/framework/src/Services/DocumentationSidebarService.php
@@ -2,10 +2,10 @@
 
 namespace Hyde\Framework\Services;
 
-use Hyde\Framework\Concerns\HasDocumentationSidebarCategories;
 use Hyde\Framework\Contracts\DocumentationSidebarServiceContract;
 use Hyde\Framework\Models\DocumentationSidebar;
 use Hyde\Framework\Models\DocumentationSidebarItem;
+use Illuminate\Support\Str;
 
 /**
  * Service class to create and manage the sidebar collection object.
@@ -15,7 +15,7 @@ use Hyde\Framework\Models\DocumentationSidebarItem;
  */
 class DocumentationSidebarService implements DocumentationSidebarServiceContract
 {
-    use HasDocumentationSidebarCategories;
+    protected array $categories = [];
 
     /**
      * The sidebar object created and managed by the service instance.
@@ -118,5 +118,56 @@ class DocumentationSidebarService implements DocumentationSidebarServiceContract
     protected function createSidebarItemFromSlug(string $slug): DocumentationSidebarItem
     {
         return DocumentationSidebarItem::parseFromFile($slug);
+    }
+
+    // Sidebar Categories
+
+    public function hasCategories(): bool
+    {
+        $this->assembleCategories();
+
+        return ! empty($this->categories);
+    }
+
+    public function getCategories(): array
+    {
+        $this->assembleCategories();
+
+        return $this->categories;
+    }
+
+    public function getItemsInCategory(string $category): DocumentationSidebar
+    {
+        return $this->sidebar->filter(function ($item) use ($category) {
+            return $item->category === Str::slug($category);
+        })->sortBy('priority')->values();
+    }
+
+    protected function assembleCategories(): void
+    {
+        foreach ($this->sidebar->sortItems() as $item) {
+            if (isset($item->category)) {
+                if (! in_array($item->category, $this->categories)) {
+                    $this->categories[] = $item->category;
+                }
+            }
+        }
+
+        if (! empty($this->categories)) {
+            $this->setCategoryOfUncategorizedItems();
+        }
+    }
+
+    protected function setCategoryOfUncategorizedItems(): void
+    {
+        foreach ($this->sidebar as $item) {
+            if (! isset($item->category)) {
+                $item->category = 'other';
+
+                if (! in_array('other', $this->categories)) {
+                    $this->categories[] = 'other';
+                }
+            }
+        }
     }
 }